Fixed a bug in FindPivot and now zeros all values under the pivot after multiplying by scalar.
There was a bug in FindPivot where it would always return the last found pivot instead of the first. This happens because we were using a parallel for without a break structure. I'm not sure how to use break in parallel fors, so I just commented the parallel for and added a break to it.